move find function to utils package
