+++ core/src/main/scala/gapt/formats/tip/util/package.scala
@@ -1,0 +1,20 @@
+package gapt.formats.tip
+
+package object util {
+
+  object find {
+    def apply[T](
+      elements: Seq[T], p: ( T ) => Boolean ): Option[( Seq[T], T, Seq[T] )] = {
+      val index = elements.indexWhere( p )
+      if ( index == -1 ) {
+        None
+      } else {
+        Some( (
+          elements.take( index ),
+          elements( index ),
+          elements.drop( index + 1 ) ) )
+      }
+    }
+  }
+
+}
